Factor saguard into its own file
